feat(router): retire `router app` and `router user` for the verbs that own them
`router app` forwarded a Market request and created the provider row a moment
before Router's own application directory would have. That head start was the
only thing it bought: the directory projects a model application into a
`providers` row whether or not anybody asked it to, which is what already
happens when the same application is installed from the Market UI. Meanwhile
the Market's own verbs know things this one never did — cloning an engine base,
picking a source, stopping and resuming — so the CLI had two install paths
where one is better.

`router user` read Router's user table, which is not the list of people on this
Olares. A row there appears the first time somebody's session reaches Router,
and the role on it is the Olares role read back rather than a second one. So
the people are `olares-cli settings users list`, and what stays here is the
lookup the verbs that take a person need: `key issue --for-user`, `quota set
--user` and `usage --user` are given a name and have to send an id.

Two private helpers survive their commands. `providerIDFromMarket` still reads
the catalog, because an application in the minutes after install is absent from
every other list while being exactly what someone names. `listUsers` and
`resolveUserID` back the four `--user` flags and the columns that put names on
UUIDs.

Router's `/console/api/market/*`, `/console/api/installed-apps` and
`/console/api/users` routes are untouched; the console front end uses all three.

Verification: go build ./..., go vet ./cmd/..., go test ./cmd/ctl/...,
gofmt -l cmd/ctl/router. Live against Router on the active profile:
`router usage summary --user yaotest004` resolves the name and answers, an
unknown name is still refused with the ones Router knows, and `router --help`
lists neither verb.

Out (deferred): an unknown word under `router` still prints help rather than
"unknown command", as it does under `market` and every other group root here.
