+++ cli/cmd/ctl/skills_commands_test.go
@@ -153,12 +153,6 @@ func TestSkillCommandPathsExist(t *testing.T) {
 		"router spec show",
 		"router spec edit",
 		"router spec restart",
-		"router app",
-		"router app catalog",
-		"router app installed",
-		"router app install",
-		"router app upgrade",
-		"router app uninstall",
 		"router call",
 		"router call chat",
 		"router call embed",
@@ -193,8 +187,6 @@ func TestSkillCommandPathsExist(t *testing.T) {
 		"router audit",
 		"router audit list",
 		"router audit get",
-		"router user",
-		"router user list",
 		"router local",
 		"router local status",
 		"router local progress",
